requirements.txt
aiofiles==24.1.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.58.0
distro==1.9.0
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
gradio==5.36.2
gradio_client==1.10.4
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
groovy==0.1.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mlflow==3.1.1
mlflow-skinny==3.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
openai==1.95.1
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==3.11.0
prometheus_client==0.22.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
ruff==0.12.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
whylabs-client==0.6.16
whylogs==1.6.4
whylogs-sketching==3.4.1.dev3
zipp==3.23.0
